requirements.txt
PyAutoGUI
opencv-python
python-dotenv
pillow

#pip install -r requirements.txt